S19-CLIP Models and Training them/requirements.txt
gradio
matplotlib
numpy
opencv-python
torch
openai-clip
tqdm
Pillow